extract struct expressions to folders based on spark grouping
